run cargo clippy
