+++ src/command/mod.rs
@@ -1,6 +1,6 @@
-pub(crate) mod command;
+pub(crate) mod action;
+pub(crate) mod application;
 pub(crate) mod handler;
 pub(crate) mod util;
 
-pub use command::*;
-pub use handler::*;
+pub use application::*;
