+++ src/main.rs
@@ -1,5 +1,4 @@
 use chrono::Utc;
-use clap::{ArgMatches, ErrorKind};
 use gluesql::{
     memory_storage::Key,
     prelude::{Glue, MemoryStorage},
@@ -7,77 +6,159 @@ use gluesql::{
 use std::collections::HashMap;
 use std::error::Error;
 use std::io::{self};
-use std::process;
 use std::sync::{Arc, Mutex};
-use tabled::{Style, TableIteratorExt};
-use tokio::task::JoinHandle;
 use tokio::time::sleep;
+use tokio::{net::UnixDatagram, sync::mpsc};
+use tokio::{sync::mpsc::Sender, task::JoinHandle};
 
-mod archived_notification;
-mod argument;
+mod command;
 mod database;
 mod notification;
 use database as db;
 mod configuration;
 mod error;
-mod input_handler;
+mod ipc;
 mod logging;
-mod notify;
+mod report;
 
-use crate::argument::{parse_arg, CLEAR, CREATE, DELETE, EXIT, HISTORY, LIST, LS, Q, QUEUE, TEST};
-use crate::configuration::{
-    generate_configuration_report, load_configuration, Configuration,
-};
+use crate::configuration::{get_configuration, Configuration};
+use crate::ipc::{create_client_uds, create_server_uds, MessageRequest, MessageResponse};
+use crate::notification::archived_notification;
+use crate::notification::notify::{notify_break, notify_work};
 use crate::notification::Notification;
-use crate::notify::{notify_break, notify_work};
+use crate::{
+    command::{handler, CommandType},
+    ipc::{get_uds_address, UdsType},
+};
 
 #[macro_use]
 extern crate log;
 
-type TaskMap = HashMap<u16, JoinHandle<()>>;
+// key: notification id, value: spawned notification task
+pub type TaskMap = HashMap<u16, JoinHandle<()>>;
 pub type ArcGlue = Arc<Mutex<Glue<Key, MemoryStorage>>>;
+pub type ArcTaskMap = Arc<Mutex<TaskMap>>;
+
+#[derive(Debug)]
+struct UserInput {
+    pub input: String,
+    // pub oneshot_tx: oneshot::Sender<String>,
+    pub source: InputSource,
+}
+
+#[derive(Debug)]
+enum InputSource {
+    StandardInput,
+    UnixDomainSocket,
+}
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
     logging::initialize_logging();
-
-    info!("info test, start pomodoro...");
     debug!("debug test, start pomodoro...");
 
-    let configuration = initialize_configuration()?;
-    let glue = initialize_db().await;
+    // TODO(young) handle result
+    let command_type = detect_command_type().await?;
 
-    let mut id_manager: u16 = 1;
-    let hash_map: Arc<Mutex<TaskMap>> = Arc::new(Mutex::new(HashMap::new()));
+    match command_type {
+        CommandType::StartUp(config) => {
+            info!("start pomodoro...");
+            debug!("CommandType::StartUp");
 
-    loop {
-        debug!("inside spawn blocking");
-        let user_input = input_handler::read_input(&mut io::stdout(), &mut io::stdin().lock());
+            let glue = initialize_db().await;
+            let mut id_manager: u16 = 1;
+            let hash_map: Arc<Mutex<TaskMap>> = Arc::new(Mutex::new(HashMap::new()));
+            let (user_input_tx, mut user_input_rx) = mpsc::channel::<UserInput>(64);
 
-        debug!("user input: {}", &user_input);
-        if let Err(e) = analyze_input(
-            &user_input,
-            &mut id_manager,
-            &hash_map,
-            &glue,
-            &configuration,
-        )
-        .await
-        {
-            println!("There was an error analyzing the input: {}", e);
-        };
+            let stdin_tx = user_input_tx.clone();
+            // TODO(young): handle tokio::spawn return value nicely so that we can use `?` inside
+            let stdinput_handle = spawn_stdinput_handler(stdin_tx);
+
+            // handle uds
+            let uds_input_tx = user_input_tx.clone();
+            let server_uds = Arc::new(create_server_uds().unwrap());
+            let server_rx = server_uds.clone();
+
+            // TODO(young): handle tokio::spawn return value nicely so that we can use `?` inside
+            let uds_input_handle = spawn_uds_input_handler(uds_input_tx, server_rx);
+
+            let server_tx = server_uds.clone();
+            // TODO(young) handle `rx.recv().await` returns None case
+            // TODO(young): handle tokio::spawn return value nicely so that we can use `?` inside
+            while let Some(user_input) = user_input_rx.recv().await {
+                // extract input
+                let input = user_input.input.as_str();
+                debug!("input: {:?}", input);
+
+                // handle input
+                match handler::user_input::handle(input, &mut id_manager, &hash_map, &glue, &config)
+                    .await
+                {
+                    Ok(output) => match user_input.source {
+                        InputSource::StandardInput => {
+                            output.iter().for_each(|o| println!("{}", o));
+                        }
+                        InputSource::UnixDomainSocket => {
+                            let client_addr = get_uds_address(UdsType::Client);
+                            server_tx
+                                .send_to(
+                                    MessageResponse::new(output).encode()?.as_slice(),
+                                    client_addr,
+                                )
+                                .await
+                                .unwrap();
+                        }
+                    },
+                    Err(e) => {
+                        println!("There was an error analyzing the input: {}", e);
+
+                        match user_input.source {
+                            InputSource::StandardInput => {}
+                            InputSource::UnixDomainSocket => {
+                                let client_addr = get_uds_address(UdsType::Client);
+                                server_tx
+                                    .send_to(
+                                        MessageResponse::new(vec![format!(
+                                            "There was an error analyzing the input: {}",
+                                            e
+                                        )])
+                                        .encode()?
+                                        .as_slice(),
+                                        client_addr,
+                                    )
+                                    .await
+                                    .unwrap();
+                            }
+                        }
+                    }
+                }
+
+                debug!("input: {:?}", user_input);
+            }
+        }
+        CommandType::UdsClient(matches) => {
+            debug!("CommandType::UdsClient");
+            let socket = create_client_uds().await?;
+
+            handler::uds_client::handle(matches, socket).await?;
+        }
     }
+
+    debug!("handle_uds_client_command called successfully");
+
+    Ok(())
 }
 
-fn initialize_configuration() -> Result<Arc<Configuration>, Box<dyn Error>> {
-    let config_matches = argument::get_config_command().get_matches();
-    let credential_file_path = config_matches.value_of("config");
+async fn detect_command_type() -> Result<CommandType, Box<dyn Error>> {
+    let matches = command::get_start_and_uds_client_command().get_matches();
+    debug!("handle_uds_client_command, matches: {:?}", &matches);
 
-    let (configuration, config_error) = load_configuration(credential_file_path)?;
-    let report = generate_configuration_report(&configuration, config_error);
-    info!("\nconfig flag result!\n{}", report);
-    
-    Ok(Arc::new(configuration))
+    let command_type = match (&matches).subcommand().is_none() {
+        true => CommandType::StartUp(get_configuration(&matches)?),
+        false => CommandType::UdsClient(matches),
+    };
+
+    Ok(command_type)
 }
 
 async fn initialize_db() -> Arc<Mutex<Glue<Key, MemoryStorage>>> {
@@ -87,233 +168,8 @@ async fn initialize_db() -> Arc<Mutex<Glue<Key, MemoryStorage>>> {
     glue
 }
 
-
-async fn analyze_input(
-    user_input: &str,
-    id_manager: &mut u16,
-    hash_map: &Arc<Mutex<TaskMap>>,
-    glue: &Arc<Mutex<Glue<Key, MemoryStorage>>>,
-    configuration: &Arc<Configuration>,
-) -> Result<(), Box<dyn Error>> {
-    let command = argument::get_command();
-    let input = user_input.split_whitespace();
-    debug!("input: {:?}", input);
-
-    let matches = match command.try_get_matches_from(input) {
-        Ok(args) => args,
-        Err(err) => {
-            match err.kind() {
-                // DisplayHelp has help message in error
-                ErrorKind::DisplayHelp => {
-                    print!("\n{}\n", err);
-                    return Ok(());
-                }
-                // clap automatically print version string with out newline.
-                ErrorKind::DisplayVersion => {
-                    println!();
-                    return Ok(());
-                }
-                _ => {
-                    print!("\n error while handling the input, {}\n", err);
-                }
-            }
-            return Err(Box::new(err));
-        }
-    };
-
-    match matches.subcommand() {
-        Some((CREATE, sub_matches)) => {
-            handle_create(sub_matches, configuration, hash_map, glue, id_manager).await?;
-        }
-        Some((QUEUE, sub_matches)) | Some((Q, sub_matches)) => {
-            handle_queue(sub_matches, configuration, hash_map, glue, id_manager).await?;
-        }
-        Some((DELETE, sub_matches)) => {
-            if sub_matches.is_present("id") {
-                // delete one
-                let id = parse_arg::<u16>(sub_matches, "id")?;
-
-                debug!("Message::Delete called! {}", id);
-
-                let result = delete_notification(id, hash_map.clone(), glue.clone()).await;
-
-                match result {
-                    Ok(_) => println!("Notification (id: {}) deleted", id),
-                    Err(e) => eprintln!("Error: {}", e),
-                };
-                debug!("Message::Delete done");
-            } else {
-                // delete all
-                debug!("Message:DeleteAll called!");
-
-                let hash_map = hash_map.lock().unwrap();
-                for (_, handle) in hash_map.iter() {
-                    handle.abort();
-                }
-                db::delete_and_archive_all_notification(glue.clone()).await;
-                println!("All Notifications deleted");
-                debug!("Message::DeleteAll done");
-            }
-        }
-        Some((LS, _)) | Some((LIST, _)) => {
-            handle_list(glue).await;
-        }
-        Some((HISTORY, _)) => {
-            handle_history(glue).await;
-        }
-        Some((TEST, _)) => {
-            handle_test(configuration).await?;
-        }
-        Some((CLEAR, _)) => {
-            print!("\x1B[2J\x1B[1;1H");
-        }
-        Some((EXIT, _)) => {
-            process::exit(0);
-        }
-        _ => (),
-    };
-    Ok(())
-}
-
-async fn handle_list(glue: &Arc<Mutex<Glue<Key, MemoryStorage>>>) {
-    debug!("Message::List called!");
-    let notifications = db::list_notification(glue.clone()).await;
-    debug!("Message::List done");
-
-    let table = notifications
-        .table()
-        .with(Style::modern().horizontal_off())
-        .to_string();
-    info!("\n{}", table);
-
-    println!("List succeed");
-}
-
-async fn handle_history(glue: &Arc<Mutex<Glue<Key, MemoryStorage>>>) {
-    debug!("Message:History called!");
-    let archived_notifications = db::list_archived_notification(glue.clone()).await;
-    debug!("Message:History done!");
-
-    let table = archived_notifications
-        .table()
-        .with(Style::modern().horizontal_off())
-        .to_string();
-    info!("\n{}", table);
-
-    println!("History succeed");
-}
-
-async fn handle_test(configuration: &Arc<Configuration>) -> Result<(), Box<dyn Error>> {
-    debug!("Message:NotificationTest called!");
-    let report = notify_work(&configuration.clone()).await?;
-    info!("\n{}", report);
-    debug!("Message:NotificationTest done");
-    println!("Notification Test called");
-
-    Ok(())
-}
-
-async fn handle_create(
-    matches: &ArgMatches,
-    configuration: &Arc<Configuration>,
-    hash_map: &Arc<Mutex<TaskMap>>,
-    glue: &ArcGlue,
-    id_manager: &mut u16,
-) -> Result<(), Box<dyn Error>> {
-    let notification = input_handler::get_new_notification(matches, id_manager, Utc::now())?;
-    let notification = match notification {
-        Some(n) => n,
-        None => return Ok(()),
-    };
-    let id = notification.get_id();
-
-    db::create_notification(glue.clone(), &notification).await;
-
-    let handle = spawn_notification(
-        configuration.clone(),
-        hash_map.clone(),
-        glue.clone(),
-        notification,
-    );
-    let mut hash_map = hash_map.lock().unwrap();
-    hash_map.insert(id, handle);
-    println!("Notification (id: {}) created", id);
-    Ok(())
-}
-
-async fn handle_queue(
-    matches: &ArgMatches,
-    configuration: &Arc<Configuration>,
-    hash_map: &Arc<Mutex<TaskMap>>,
-    glue: &ArcGlue,
-    id_manager: &mut u16,
-) -> Result<(), Box<dyn Error>> {
-    let last_expired_notification = db::read_last_expired_notification(glue.clone()).await;
-
-    let created_at = match last_expired_notification {
-        Some(n) => {
-            debug!("last_expired_notification: {:?}", &n);
-
-            let (_, _, _, _, _, work_expired_at, break_expired_at) = n.get_values();
-
-            work_expired_at.max(break_expired_at)
-        }
-        None => Utc::now(),
-    };
-
-    let notification = input_handler::get_new_notification(matches, id_manager, created_at)?;
-    let notification = match notification {
-        Some(n) => n,
-        None => return Ok(()),
-    };
-    let id = notification.get_id();
-    db::create_notification(glue.clone(), &notification).await;
-
-    let handle = spawn_notification(
-        configuration.clone(),
-        hash_map.clone(),
-        glue.clone(),
-        notification,
-    );
-    let mut hash_map = hash_map.lock().unwrap();
-    hash_map.insert(id, handle);
-    println!("Notification (id: {}) created and queued", id);
-    Ok(())
-}
-
-async fn delete_notification(
-    id: u16,
-    hash_map: Arc<Mutex<TaskMap>>,
-    glue: ArcGlue,
-) -> Result<(), Box<dyn Error>> {
-    let notification = db::read_notification(glue.clone(), id).await;
-    if notification.is_none() {
-        return Err(format!(
-            "deleting id ({}) failed. Corresponding notification does not exist",
-            id
-        )
-        .into());
-    }
-
-    {
-        let mut hash_map = hash_map.lock().unwrap();
-
-        hash_map
-            .get(&id)
-            .ok_or(format!("failed to corresponding task (id: {})", &id))?
-            .abort();
-
-        hash_map
-            .remove(&id)
-            .ok_or(format!("failed to remove id ({})", id))?;
-    }
-
-    db::delete_and_archive_notification(glue, id).await;
-
-    Ok(())
-}
-
-fn spawn_notification(
+// TODO(young): refactor and move to proper place
+pub fn spawn_notification(
     configuration: Arc<Configuration>,
     hash_map: Arc<Mutex<TaskMap>>,
     glue: ArcGlue,
@@ -338,7 +194,7 @@ fn spawn_notification(
             if let Ok(report) = result {
                 info!("\n{}", report);
                 println!("Notification report generated");
-                input_handler::write_output(&mut io::stdout());
+                command::util::write_output(&mut io::stdout());
             }
         }
 
@@ -352,15 +208,64 @@ fn spawn_notification(
             if let Ok(report) = result {
                 info!("\n{}", report);
                 println!("Notification report generated");
-                input_handler::write_output(&mut io::stdout());
+                command::util::write_output(&mut io::stdout());
             }
         }
 
-        let result = delete_notification(id, hash_map, glue.clone()).await;
+        let result = notification::delete_notification(id, hash_map, glue.clone()).await;
         if result.is_err() {
             trace!("error occurred while deleting notification");
         }
 
         debug!("id: {}, notification work time done!", id);
+    })
+}
+
+fn spawn_stdinput_handler(tx: Sender<UserInput>) -> JoinHandle<()> {
+    tokio::spawn(async move {
+        loop {
+            debug!("inside stdin task");
+            let user_input = command::util::read_input(&mut io::stdout(), &mut io::stdin().lock());
+            debug!("user input: {}", &user_input);
+
+            let _ = tx
+                .send(UserInput {
+                    input: user_input,
+                    source: InputSource::StandardInput,
+                })
+                .await;
+        }
+    })
+}
+
+fn spawn_uds_input_handler(
+    uds_tx: Sender<UserInput>,
+    server_rx: Arc<UnixDatagram>,
+) -> JoinHandle<()> {
+    tokio::spawn(async move {
+        // TODO(young) handle result
+        let rx = server_rx;
+        let mut buf = vec![0u8; 256];
+        debug!("rx is initialized successfully");
+        loop {
+            debug!("inside unix domain socket task");
+            // TODO(young) handle result
+            let (size, addr) = rx.recv_from(&mut buf).await.unwrap();
+            debug!("size: {:?}, addr: {:?}", size, addr);
+
+            if let Some(path) = addr.as_pathname() {
+                // ignore request from unnamed address
+                if path != get_uds_address(ipc::UdsType::Client).as_path() {
+                    debug!("addr is different");
+                    continue;
+                }
+            }
+
+            let user_input: UserInput =
+                MessageRequest::into(MessageRequest::decode(&buf[..size]).unwrap());
+            debug!("user_input: {:?}", user_input);
+
+            let _ = uds_tx.send(user_input).await.unwrap();
+        }
     })
 }
